requirements.txt
streamlit
google-generativeai
requests
beautifulsoup4
pandas
st-gsheets-connection
st_paywall
hmac